revamp(frontend): move Header component to components directory
Moved Header.tsx from assets to components and updated imports in App.tsx and Home.tsx to reflect the new location. This improves project structure by placing components in the appropriate directory.
